refactor: split comb path state types
