+++ crates/celox/src/logic_tree/comb.rs
@@ -1,3 +1,9 @@
+mod path;
+mod state;
+
+pub use path::{LogicPath, LogicPathTarget};
+pub use state::{BoundaryMap, SymbolicStore};
+
 use std::{collections::BTreeSet, fmt, hash::Hash};
 
 use serde::{Deserialize, Serialize};
@@ -9,8 +15,7 @@ use crate::parser::{LoweringPhase, resolve_total_width};
 use crate::{
     HashMap, HashSet,
     ir::{
-        BinaryOp, BitAccess, CombObserver, LogicPathId, RuntimeEventKind, RuntimeEventSite,
-        UnaryOp, VarAtomBase,
+        BinaryOp, BitAccess, CombObserver, RuntimeEventKind, RuntimeEventSite, UnaryOp, VarAtomBase,
     },
     parser::bitaccess::{celox_value_from_comptime, eval_constexpr, eval_var_select},
 };
@@ -23,33 +28,7 @@ use veryl_analyzer::ir::{
 };
 use veryl_analyzer::value::{Value, byte_value_to_string};
 
-// SymbolicStore: Maps variable IDs to their current symbolic representation.
-// Each variable is managed by a RangeStore, which tracks bit-ranges and their associated SLT nodes.
-pub type SymbolicStore<A> = HashMap<VarId, RangeStore<Option<(NodeId, HashSet<VarAtomBase<A>>)>>>;
-pub type BoundaryMap<A> = HashMap<A, BTreeSet<usize>>;
-
-#[derive(Clone)]
-struct LoopControlState {
-    store: SymbolicStore<VarId>,
-    boundaries: BoundaryMap<VarId>,
-    continue_expr: NodeId,
-    continue_sources: HashSet<VarAtomBase<VarId>>,
-}
-
-#[derive(Clone)]
-struct FunctionControlState {
-    store: SymbolicStore<VarId>,
-    boundaries: BoundaryMap<VarId>,
-    live_expr: NodeId,
-    live_sources: HashSet<VarAtomBase<VarId>>,
-}
-
-#[derive(Clone)]
-struct FunctionLoopControlState {
-    function: FunctionControlState,
-    continue_expr: NodeId,
-    continue_sources: HashSet<VarAtomBase<VarId>>,
-}
+use state::{FunctionControlState, FunctionLoopControlState, LoopControlState};
 
 #[derive(Debug, Clone, Copy, PartialEq, Eq, PartialOrd, Ord, Hash, Serialize, Deserialize)]
 pub struct NodeId(pub usize);
@@ -983,151 +962,6 @@ impl<A: fmt::Debug + fmt::Display + Hash + Eq + Clone> SLTNode<A> {
                 writeln!(f, "{}Slice[{}:{}]", indent, access.lsb, access.msb)?;
                 arena.get(*expr).fmt_recursive(f, depth + 1, arena)
             }
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(bound(
-    serialize = "A: Serialize + std::hash::Hash + Eq",
-    deserialize = "A: Deserialize<'de> + std::hash::Hash + Eq + Clone"
-))]
-pub enum LogicPathTarget<A: Hash + Eq + Clone> {
-    Var(VarAtomBase<A>),
-    CombCaptureEvent {
-        site_id: u32,
-        guard: Option<NodeId>,
-        emit_on_true: bool,
-        args: Vec<NodeId>,
-        loop_runner: Option<NodeId>,
-        fatal_error_code: Option<i64>,
-        consume_enabled: bool,
-    },
-}
-
-impl<A: Hash + Eq + Clone> LogicPathTarget<A> {
-    pub fn var(&self) -> Option<&VarAtomBase<A>> {
-        match self {
-            LogicPathTarget::Var(var) => Some(var),
-            LogicPathTarget::CombCaptureEvent { .. } => None,
-        }
-    }
-}
-
-impl<A: fmt::Display + Hash + Eq + Clone> fmt::Display for LogicPathTarget<A> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        match self {
-            LogicPathTarget::Var(var) => write!(f, "{var}"),
-            LogicPathTarget::CombCaptureEvent { site_id, .. } => {
-                write!(f, "capture_event({site_id})")
-            }
-        }
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq, Serialize, Deserialize)]
-#[serde(bound(
-    serialize = "A: Serialize + std::hash::Hash + Eq",
-    deserialize = "A: Deserialize<'de> + std::hash::Hash + Eq + Clone"
-))]
-pub struct LogicPath<A: Hash + Eq + Clone> {
-    pub target: LogicPathTarget<A>,
-    pub sources: HashSet<VarAtomBase<A>>,
-    pub local_inputs: Vec<(A, NodeId)>,
-    pub order_before: HashSet<LogicPathId>,
-    pub comb_capture_enable_sites: Vec<u32>,
-    pub pre_lower_nodes: Vec<NodeId>,
-    pub expr: NodeId,
-}
-
-impl<A: fmt::Display + Hash + Eq + Clone> fmt::Display for LogicPath<A> {
-    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
-        write!(f, "{}", self.target)
-    }
-}
-
-impl<A: fmt::Debug + fmt::Display + Hash + Eq + Clone> LogicPath<A> {
-    pub fn map_addr<B: Hash + Eq + Clone, F>(
-        &self,
-        arena: &SLTNodeArena<A>,
-        target_arena: &mut SLTNodeArena<B>,
-        cache: &mut HashMap<NodeId, NodeId>,
-        f: &F,
-    ) -> LogicPath<B>
-    where
-        F: Fn(&A) -> B,
-    {
-        LogicPath {
-            target: match &self.target {
-                LogicPathTarget::Var(var) => LogicPathTarget::Var(VarAtomBase::new(
-                    f(&var.id),
-                    var.access.lsb,
-                    var.access.msb,
-                )),
-                LogicPathTarget::CombCaptureEvent {
-                    site_id,
-                    guard,
-                    emit_on_true,
-                    args,
-                    loop_runner,
-                    fatal_error_code,
-                    consume_enabled,
-                } => LogicPathTarget::CombCaptureEvent {
-                    site_id: *site_id,
-                    guard: guard.map(|node| {
-                        arena
-                            .get(node)
-                            .map_addr(node, arena, target_arena, cache, f)
-                    }),
-                    emit_on_true: *emit_on_true,
-                    args: args
-                        .iter()
-                        .map(|node| {
-                            arena
-                                .get(*node)
-                                .map_addr(*node, arena, target_arena, cache, f)
-                        })
-                        .collect(),
-                    loop_runner: loop_runner.map(|node| {
-                        arena
-                            .get(node)
-                            .map_addr(node, arena, target_arena, cache, f)
-                    }),
-                    fatal_error_code: *fatal_error_code,
-                    consume_enabled: *consume_enabled,
-                },
-            },
-            sources: self
-                .sources
-                .iter()
-                .map(|v| VarAtomBase::new(f(&v.id), v.access.lsb, v.access.msb))
-                .collect(),
-            local_inputs: self
-                .local_inputs
-                .iter()
-                .map(|(id, node)| {
-                    (
-                        f(id),
-                        arena
-                            .get(*node)
-                            .map_addr(*node, arena, target_arena, cache, f),
-                    )
-                })
-                .collect(),
-            order_before: self.order_before.clone(),
-            comb_capture_enable_sites: self.comb_capture_enable_sites.clone(),
-            pre_lower_nodes: self
-                .pre_lower_nodes
-                .iter()
-                .map(|node| {
-                    arena
-                        .get(*node)
-                        .map_addr(*node, arena, target_arena, cache, f)
-                })
-                .collect(),
-            expr: arena
-                .get(self.expr)
-                .map_addr(self.expr, arena, target_arena, cache, f),
         }
     }
 }
